Add dynamic params configuration and move Chat API
