+++ .env
@@ -1,1 +1,1 @@
-VITE_CHATGPT_KEY=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
+VITE_CHATGPT_KEY=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
